refactor: relocate session-lifecycle helpers into session-lifecycle/ (R3)
